wip: add basic support for sat geometry
